pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "git-analytics-cli"
version = "0.1.2"
description = "Local-first Git habit analytics — scan your repos, get a developer persona and health report"
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "Git Analytics Contributors" }]
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = ["Pillow>=8.0"]

[project.urls]
Homepage = "https://github.com/mark-618/git-analytics"
Repository = "https://github.com/mark-618/git-analytics"
Issues = "https://github.com/mark-618/git-analytics/issues"

[project.scripts]
git-analytics = "run:main"

[tool.setuptools]
py-modules = ["run", "git_analytics", "generate_report", "gen_share_card"]

[tool.setuptools.data-files]
"share/git-analytics" = ["share-card.html"]
